allow user to specify a different storage pool
Some default installs of Fedora do not create enough space for 40G of VM
images to be written to '/var/lib/libvirt/images'. Allow a user to
specify a different libvirt storage pool, via a question prompt, to be
used when provisioning beaker VMs. This storage pool is verified to be
at least 40G in total size and that the pool is configured to autostart
and is online.

It is currently an exercise for the user to create this new storage
pool, though this could be solved with a pre-task.

Change-Id: I3e865e57d6f0f64faf251c2d06f1fea2391a20fd
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>